+++ sdk/python/pulumi_rancher2/project_logging.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from . import _utilities, _tables
 from . import outputs
 from ._inputs import *
@@ -27,7 +27,7 @@ class ProjectLogging(pulumi.CustomResource):
                  labels: Optional[pulumi.Input[Mapping[str, Any]]] = None,
                  name: Optional[pulumi.Input[str]] = None,
                  namespace_id: Optional[pulumi.Input[str]] = None,
-                 output_flush_interval: Optional[pulumi.Input[float]] = None,
+                 output_flush_interval: Optional[pulumi.Input[int]] = None,
                  output_tags: Optional[pulumi.Input[Mapping[str, Any]]] = None,
                  project_id: Optional[pulumi.Input[str]] = None,
                  splunk_config: Optional[pulumi.Input[pulumi.InputType['ProjectLoggingSplunkConfigArgs']]] = None,
@@ -68,7 +68,7 @@ class ProjectLogging(pulumi.CustomResource):
         :param pulumi.Input[Mapping[str, Any]] labels: Labels for Project Logging object (map)
         :param pulumi.Input[str] name: The name of the Project Logging config (string)
         :param pulumi.Input[str] namespace_id: The namespace id from Project logging (string)
-        :param pulumi.Input[float] output_flush_interval: How often buffered logs would be flushed. Default: `3` seconds (int)
+        :param pulumi.Input[int] output_flush_interval: How often buffered logs would be flushed. Default: `3` seconds (int)
         :param pulumi.Input[Mapping[str, Any]] output_tags: The output tags for Project Logging (map)
         :param pulumi.Input[str] project_id: The project id to configure logging (string)
         :param pulumi.Input[pulumi.InputType['ProjectLoggingSplunkConfigArgs']] splunk_config: The splunk config for Project Logging. For `kind = splunk`. Conflicts with `custom_target_config`, `elasticsearch_config`, `fluentd_config`, `kafka_config`, and `syslog_config` (list maxitems:1)
@@ -130,7 +130,7 @@ class ProjectLogging(pulumi.CustomResource):
             labels: Optional[pulumi.Input[Mapping[str, Any]]] = None,
             name: Optional[pulumi.Input[str]] = None,
             namespace_id: Optional[pulumi.Input[str]] = None,
-            output_flush_interval: Optional[pulumi.Input[float]] = None,
+            output_flush_interval: Optional[pulumi.Input[int]] = None,
             output_tags: Optional[pulumi.Input[Mapping[str, Any]]] = None,
             project_id: Optional[pulumi.Input[str]] = None,
             splunk_config: Optional[pulumi.Input[pulumi.InputType['ProjectLoggingSplunkConfigArgs']]] = None,
@@ -152,7 +152,7 @@ class ProjectLogging(pulumi.CustomResource):
         :param pulumi.Input[Mapping[str, Any]] labels: Labels for Project Logging object (map)
         :param pulumi.Input[str] name: The name of the Project Logging config (string)
         :param pulumi.Input[str] namespace_id: The namespace id from Project logging (string)
-        :param pulumi.Input[float] output_flush_interval: How often buffered logs would be flushed. Default: `3` seconds (int)
+        :param pulumi.Input[int] output_flush_interval: How often buffered logs would be flushed. Default: `3` seconds (int)
         :param pulumi.Input[Mapping[str, Any]] output_tags: The output tags for Project Logging (map)
         :param pulumi.Input[str] project_id: The project id to configure logging (string)
         :param pulumi.Input[pulumi.InputType['ProjectLoggingSplunkConfigArgs']] splunk_config: The splunk config for Project Logging. For `kind = splunk`. Conflicts with `custom_target_config`, `elasticsearch_config`, `fluentd_config`, `kafka_config`, and `syslog_config` (list maxitems:1)
@@ -261,7 +261,7 @@ class ProjectLogging(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="outputFlushInterval")
-    def output_flush_interval(self) -> pulumi.Output[Optional[float]]:
+    def output_flush_interval(self) -> pulumi.Output[Optional[int]]:
         """
         How often buffered logs would be flushed. Default: `3` seconds (int)
         """
